+++ LoveDialogue/LoveDialogueParser.lua
@@ -3,6 +3,7 @@ local LD_PATH = (...):match('(.-)[^%.]+$')
 local CallbackHandler = require(LD_PATH .. "CallbackHandler")
 local PortraitManager = require(LD_PATH .. "PortraitManager")
 local LD_Character = require(LD_PATH .. "LD_Character")
+local CharacterParser = require(LD_PATH .. "LD_CharacterParser")
 
 local function loadLuaFile(filePath)
     local chunk, err = loadfile(filePath)
